[refactor] Refactor BERT script folder (#744)
* move model definition to model/

* update tutorial

* add squad test

* add test argument

* refactor data folder

* fix tutorial

* fix lint

* fix import

* fix import

* update index.rst

* fix tutorial
