+++ scripts/bert/data/__init__.py
@@ -18,5 +18,5 @@
 # under the License.
 
 # pylint: disable=wildcard-import
-"""BERT Module."""
-from . import model, data
+"""BERT data."""
+from . import qa, classification, embedding, transform, ner
