+++ scripts/bert/embedding.py
@@ -14,9 +14,9 @@ from gluonnlp.data import BERTTokenizer, BERTSentenceTransform
 from gluonnlp.base import get_home_dir
 
 try:
-    from dataset import BertEmbeddingDataset
+    from data.embedding import BertEmbeddingDataset
 except ImportError:
-    from .dataset import BertEmbeddingDataset
+    from .data.embedding import BertEmbeddingDataset
 
 try:
     unicode
